moved router into its own component
