+++ elements/outline-player/analysis.json
@@ -738,16 +738,40 @@
           }
         },
         {
+          "name": "_haxcmsRouterManifestSubscribeHandler",
+          "description": "",
+          "privacy": "protected",
+          "sourceRange": {
+            "start": {
+              "line": 429,
+              "column": 2
+            },
+            "end": {
+              "line": 432,
+              "column": 3
+            }
+          },
+          "metadata": {},
+          "params": [
+            {
+              "name": "e"
+            }
+          ],
+          "return": {
+            "type": "void"
+          }
+        },
+        {
           "name": "refreshDynamicPositions",
           "description": "Update calculations for fill-remaining and arrow position",
           "privacy": "public",
           "sourceRange": {
             "start": {
-              "line": 432,
+              "line": 437,
               "column": 2
             },
             "end": {
-              "line": 448,
+              "line": 453,
               "column": 3
             }
           },
@@ -763,11 +787,11 @@
           "privacy": "protected",
           "sourceRange": {
             "start": {
-              "line": 453,
+              "line": 458,
               "column": 2
             },
             "end": {
-              "line": 466,
+              "line": 471,
               "column": 3
             }
           },
@@ -787,11 +811,11 @@
           "privacy": "public",
           "sourceRange": {
             "start": {
-              "line": 471,
+              "line": 476,
               "column": 2
             },
             "end": {
-              "line": 488,
+              "line": 493,
               "column": 3
             }
           },
@@ -815,11 +839,11 @@
           "privacy": "protected",
           "sourceRange": {
             "start": {
-              "line": 493,
+              "line": 498,
               "column": 2
             },
             "end": {
-              "line": 501,
+              "line": 506,
               "column": 3
             }
           },
@@ -842,11 +866,11 @@
           "privacy": "public",
           "sourceRange": {
             "start": {
-              "line": 506,
+              "line": 511,
               "column": 2
             },
             "end": {
-              "line": 511,
+              "line": 516,
               "column": 3
             }
           },
@@ -863,11 +887,11 @@
           "privacy": "public",
           "sourceRange": {
             "start": {
-              "line": 516,
+              "line": 521,
               "column": 2
             },
             "end": {
-              "line": 521,
+              "line": 526,
               "column": 3
             }
           },
@@ -884,11 +908,11 @@
           "privacy": "public",
           "sourceRange": {
             "start": {
-              "line": 526,
+              "line": 531,
               "column": 2
             },
             "end": {
-              "line": 528,
+              "line": 533,
               "column": 3
             }
           },
@@ -908,11 +932,11 @@
           "privacy": "public",
           "sourceRange": {
             "start": {
-              "line": 533,
+              "line": 538,
               "column": 2
             },
             "end": {
-              "line": 535,
+              "line": 540,
               "column": 3
             }
           },
@@ -932,11 +956,11 @@
           "privacy": "public",
           "sourceRange": {
             "start": {
-              "line": 540,
+              "line": 545,
               "column": 2
             },
             "end": {
-              "line": 554,
+              "line": 559,
               "column": 3
             }
           },
@@ -956,11 +980,11 @@
           "privacy": "protected",
           "sourceRange": {
             "start": {
-              "line": 559,
+              "line": 564,
               "column": 2
             },
             "end": {
-              "line": 573,
+              "line": 578,
               "column": 3
             }
           },
@@ -983,11 +1007,11 @@
           "privacy": "private",
           "sourceRange": {
             "start": {
-              "line": 578,
+              "line": 583,
               "column": 2
             },
             "end": {
-              "line": 580,
+              "line": 585,
               "column": 3
             }
           },
@@ -997,79 +1021,6 @@
               "name": "outlineTitle"
             }
           ]
-        },
-        {
-          "name": "__mapMenuLinkClickedHandler",
-          "description": "",
-          "privacy": "private",
-          "sourceRange": {
-            "start": {
-              "line": 582,
-              "column": 2
-            },
-            "end": {
-              "line": 584,
-              "column": 3
-            }
-          },
-          "metadata": {},
-          "params": [
-            {
-              "name": "e"
-            }
-          ],
-          "return": {
-            "type": "void"
-          }
-        },
-        {
-          "name": "_haxcmsRouterManifestInit",
-          "description": "",
-          "privacy": "protected",
-          "sourceRange": {
-            "start": {
-              "line": 586,
-              "column": 2
-            },
-            "end": {
-              "line": 591,
-              "column": 3
-            }
-          },
-          "metadata": {},
-          "params": [
-            {
-              "name": "manifest"
-            }
-          ],
-          "return": {
-            "type": "void"
-          }
-        },
-        {
-          "name": "_haxcmsRouterManifestChanged",
-          "description": "Event handler for 'haxcms-router-manifest-changed'",
-          "privacy": "protected",
-          "sourceRange": {
-            "start": {
-              "line": 597,
-              "column": 2
-            },
-            "end": {
-              "line": 601,
-              "column": 3
-            }
-          },
-          "metadata": {},
-          "params": [
-            {
-              "name": "e",
-              "type": "event"
-            }
-          ],
-          "return": {
-            "type": "void"
-          }
         }
       ],
       "staticMethods": [],
@@ -1081,7 +1032,7 @@
           "column": 28
         },
         "end": {
-          "line": 602,
+          "line": 586,
           "column": 1
         }
       },
